+++ src/step/content/replace/DomReplaceCommand.ts
@@ -15,19 +15,25 @@ export abstract class DomReplaceCommand<T extends HTMLElement = HTMLElement, C e
     const replacer = await this.createReplacer(context)
     do {
       contents = result
-      const doc = inputFile.dom.window.document.documentElement
+      const doc = inputFile.document.documentElement
       const elements = doc.querySelectorAll<T>(this.selector)
       if (elements.length > 0) {
         for (const element of elements) {
           const replaced = await replacer.replace(element)
           element.replaceWith(replaced)
         }
-        result = inputFile.dom.serialize()
+        result = inputFile.serialize()
       }
     } while (result != contents)
     inputFile.contents = result
     return inputFile
   }
 
+  /**
+   * Creates a replacer in a given context.
+   *
+   * @param context
+   * @protected
+   */
   protected abstract createReplacer(context: C): Promise<DomReplacer<T>>
 }