+++ src/step/content/replace/html/tag/HtmlTagReplaceCommand.ts
@@ -1,24 +1,18 @@
-import {RegexReplaceCommand} from "../../RegexReplaceCommand"
 import {ReplacerFactory} from "../../ReplacerFactory"
-import {RegexReplacer} from "../../RegexReplacer"
-import {SsgContext} from "../../../../../SsgContext"
+import {DomReplaceCommand} from "../../DomReplaceCommand"
+import {DomReplacer} from "../../DomReplacer"
+import {HtmlSsgContext} from "../../../../../HtmlSsgContext"
 
 /**
- * A command that will replace tag expressions (`<tag-name>content</tag-name>`).
- *
- * @deprecated This cannot be reliable using Regex and will only work for non-nested tags cases. Use a
- * DomReplaceCommand instead.
+ * Command to replace HTML tags by name.
  */
-export class HtmlTagReplaceCommand extends RegexReplaceCommand {
-  /**
-   * @param tagName The name of the tag to look for ("tag-name" for <tag-name>).
-   * @param replacerFactory The delegate to create replacers with the current context.
-   */
-  constructor(protected tagName: string, protected replacerFactory: ReplacerFactory<RegexReplacer>) {
-    super(new RegExp(`<${tagName}(\\s+.*?)?>\\s*(.+?)\\s*</${tagName}>`, "gm"))
+export class HtmlTagReplaceCommand<T extends HTMLElement = HTMLElement> extends DomReplaceCommand<T> {
+
+  constructor(protected tagName: string, protected replacerFactory: ReplacerFactory<DomReplacer<T>>) {
+    super(tagName)
   }
 
-  protected createReplacer(context: SsgContext): Promise<RegexReplacer> {
+  protected createReplacer(context: HtmlSsgContext): Promise<DomReplacer<T>> {
     return this.replacerFactory.create(context)
   }
 }