+++ src/step/content/replace/html/tag/index.ts
@@ -1,1 +1,2 @@
+export * from "./RegexHtmlTagReplaceCommand"
 export * from "./HtmlTagReplaceCommand"